+++ sdk/src/ast/file.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Serialize, Deserialize)]
 pub struct File {
-    pub id: u128,
+    pub id: u32,
     pub children: RefCell<Vec<FileChildType>>,
     pub name: String,
     pub path: String,
